LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/access-to-coop-page-from-other_coop-ro_cross-origin.https-expected.txt


FAIL access-to-coop-page-from-other (COOP-RO) assert_equals: expected (string) "coop" but got (undefined) undefined
